move implementations to appropriate place and name
