+++ src/main/java/com/example/loborems/Services/DOAInteractionImpl.java
@@ -1,18 +1,19 @@
-package com.example.loborems.interfaces;
+package com.example.loborems.services;
 
 import java.util.List;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 
+import com.example.loborems.interfaces.DOA;
 import com.example.loborems.models.Interaction;
 import com.example.loborems.util.HibernateUtil;
 
-public class DOAInteraction implements DOA<Interaction> {
+public class DOAInteractionImpl implements DOA<Interaction> {
 
     private final SessionFactory sessionFactory;
 
-    public DOAInteraction() {
+    public DOAInteractionImpl() {
         this.sessionFactory = HibernateUtil.getSessionFactory();
     }
